requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.5
requests==2.31.0
httpx==0.24.1
python-dotenv==1.0.0
supabase==2.3.4
websockets==11.0.3
